project/pipe/requirements.txt
python-dotenv
pika
pillow
opencv-python
numpy
datetime
tqdm
psutil
tabulate
requests
asyncio
